show-tree is not a config command anymore
