+++ src/statue/cli/config/__init__.py
@@ -13,12 +13,13 @@ from statue.cli.config.config_contexts import (
 )
 from statue.cli.config.config_general import set_history_size_cli, set_mode_cli
 from statue.cli.config.config_init import init_config_cli
-from statue.cli.config.config_show import show_config_cli, show_config_tree_cli
+from statue.cli.config.config_show import show_config_cli
 from statue.cli.config.config_sources import (
     add_source_to_configuration_cli,
     edit_source_in_configuration_cli,
     remove_source_from_configuration_cli,
 )
+from statue.cli.show_tree import show_tree_cli
 
 __all__ = [
     "config_cli",
@@ -36,5 +37,5 @@ __all__ = [
     "remove_source_from_configuration_cli",
     "init_config_cli",
     "show_config_cli",
-    "show_config_tree_cli",
+    "show_tree_cli",
 ]
